requirements.txt
# Cài đặt dependencies chung
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-Migrate>=3.0.1
Flask-Login>=0.5.0
Flask-WTF>=0.15.1
Werkzeug>=2.0.1
psycopg2-binary>=2.9.1
email-validator>=1.1.3
requests>=2.26.0
beautifulsoup4>=4.10.0
playwright>=1.25.0
python-dotenv>=0.19.0

# NLP và phân tích
nltk>=3.6.3
numpy>=1.21.2
pandas>=1.3.3
scipy>=1.7.1
scikit-learn>=1.0
gensim>=4.1.2
joblib>=1.1.0
tqdm>=4.64.0

# Visualization
plotly>=5.3.1
